Add Post form validation & saving of data through database Add search single post by editing Change validation for mobile integer to numeric Add formik validations for post form Add post api routes

+++ src/app/Http/Controllers/UserApiController.php
@@ -45,7 +45,7 @@ class UserApiController extends Controller
             'password' => 'required',
             'userType' => 'required',
             'emailAddress' => 'required',
-            'mobile' => 'integer',
+            'mobile' => 'numeric',
         ]);
 
         if ($validator->fails()) {
